requirements.txt
numpy
opencv-python
pillow
matplotlib
scipy
scikit-image
